+++ src/MythicScribe.ts
@@ -16,6 +16,7 @@ import {
 } from './common/datasets/datasets';
 import { configHandler } from './common/utils/configutils';
 import { scribeColorProvider } from './common/color/colorprovider';
+import { MythicNodeHandler } from './common/mythicnodes/MythicNode';
 
 export let ctx: vscode.ExtensionContext;
 
@@ -63,6 +64,8 @@ export async function activate(context: vscode.ExtensionContext) {
     if (activeEditor) {
         SubscriptionHelper.updateSubscriptions(activeEditor.document);
     }
+
+    MythicNodeHandler.scanAllDocuments();
 }
 
 export function deactivate() {}